Move BufferRanges to separate file
